+++ src/Core/Data/Dto/Message/MessageConnectionDto.cs
@@ -2,7 +2,7 @@ namespace GamaEdtech.Data.Dto.Message
 {
     public sealed class MessageConnectionDto
     {
-        public int ConnectionId { get; set; }
+        public long ConnectionId { get; set; }
         public string? ConnectionName { get; set; }
         public int UnreadCount { get; set; }
     }
