+++ src/Presentation/Api/Areas/Admin/Controllers/BlogsController.cs
@@ -53,7 +53,7 @@ namespace GamaEdtech.Presentation.Api.Areas.Admin.Controllers
                     var userIds = await identityService.Value.GetUserIdsAsync(new EmailEqualsSpecification(request.Email));
                     if (userIds.Data?.Count > 0)
                     {
-                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, int>(userIds.Data));
+                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, long>(userIds.Data));
                     }
                 }
 
@@ -62,7 +62,7 @@ namespace GamaEdtech.Presentation.Api.Areas.Admin.Controllers
                     var userIds = await identityService.Value.GetUserIdsAsync(new UsernameEqualsSpecification(request.Username));
                     if (userIds.Data?.Count > 0)
                     {
-                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, int>(userIds.Data));
+                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, long>(userIds.Data));
                     }
                 }
 
@@ -272,7 +272,7 @@ namespace GamaEdtech.Presentation.Api.Areas.Admin.Controllers
                     var userIds = await identityService.Value.GetUserIdsAsync(new NameContainsSpecification(request.CommenterName));
                     if (userIds.Data?.Count > 0)
                     {
-                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, int>(userIds.Data));
+                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, long>(userIds.Data));
                     }
                 }
 
@@ -281,7 +281,7 @@ namespace GamaEdtech.Presentation.Api.Areas.Admin.Controllers
                     var userIds = await identityService.Value.GetUserIdsAsync(new EmailEqualsSpecification(request.CommenterEmail));
                     if (userIds.Data?.Count > 0)
                     {
-                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, int>(userIds.Data));
+                        specification = specification.And(new CreationUserIdContainsSpecification<Contribution, ApplicationUser, long>(userIds.Data));
                     }
                 }
